Backend/speech_service/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
cryptography==46.0.7
fastapi==0.135.3
flatbuffers==25.12.19
google-api-core==2.30.2
google-auth==2.49.1
google-cloud-speech==2.38.0
google-cloud-texttospeech==2.36.0
googleapis-common-protos==1.74.0
grpcio==1.80.0
grpcio-status==1.80.0
h11==0.16.0
idna==3.11
mpmath==1.3.0
numpy==2.4.4
onnxruntime==1.24.4
packaging==26.1
pathvalidate==3.3.1
piper-tts==1.4.2
proto-plus==1.27.2
protobuf==6.33.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
python-multipart==0.0.24
requests==2.33.1
srt==3.5.3
starlette==1.0.0
sympy==1.14.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uvicorn==0.44.0
vosk==0.3.45
websockets==16.0
